requirements.txt
python-telegram-bot==20.7
dnspython
requests
colorama